vmci: type promotion bug in qp_host_get_user_memory()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-408.el8
commit-author Dan Carpenter <dan.carpenter@oracle.com>
commit 7fb2fd4e25fc1fb10dcb30b5519de257cfeae84c

The problem is that if get_user_pages_fast() fails and returns a
negative error code, it gets type promoted to a high positive value and
treated as a success.

Fixes: 06164d2b72aa ("VMCI: queue pairs implementation.")
(cherry picked from commit 7fb2fd4e25fc1fb10dcb30b5519de257cfeae84c)

+++ drivers/misc/vmw_vmci/vmci_queue_pair.c
@@ -662,7 +662,7 @@ static int qp_host_get_user_memory(u64 produce_uva,
 				     produce_q->kernel_if->num_pages,
 				     FOLL_WRITE,
 				     produce_q->kernel_if->u.h.header_page);
-	if (retval < produce_q->kernel_if->num_pages) {
+	if (retval < (int)produce_q->kernel_if->num_pages) {
 		pr_debug("get_user_pages_fast(produce) failed (retval=%d)",
 			retval);
 		qp_release_pages(produce_q->kernel_if->u.h.header_page,
@@ -675,7 +675,7 @@ static int qp_host_get_user_memory(u64 produce_uva,
 				     consume_q->kernel_if->num_pages,
 				     FOLL_WRITE,
 				     consume_q->kernel_if->u.h.header_page);
-	if (retval < consume_q->kernel_if->num_pages) {
+	if (retval < (int)consume_q->kernel_if->num_pages) {
 		pr_debug("get_user_pages_fast(consume) failed (retval=%d)",
 			retval);
 		qp_release_pages(consume_q->kernel_if->u.h.header_page,
